Remove homebrew
Not impressed; either homebrew works differently on MacOS, or Mac users like it because they don't have any real package managers.
It's basically accumulating another linux distro in it's /home/linuxbrew/ folder, which it adds to your path.
As soon as you ask it to install something that has a python dependency, it'll start installing versions of python in there and pre-pending them to your path.
It doesn't take long before this starts to break python stuff.